Model::personContainsKeywordsPredicate: Renamed file to suit find command. Added switch case to find proper person with prefix.
